fix: removed useless comments and changed the structure
